requirements.txt
django==2.2.6
pillow==7.0.0
sorl-thumbnail==12.6.3
Unidecode==1.1.1 
djangorestframework==3.12.2
